changed data type from main only

+++ src/main/java/org/example/Calculator.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Calculator {
 
-    public void display(int a, int b) {
+    public void display(float a,float b) {
 
         Addition addition = new Addition();
         Subtraction subtraction = new Subtraction();
